Refactor session event and elicitation handling
