+++ src/CodexAcpServer.ts
@@ -6,7 +6,6 @@ import {
     type SessionModeState
 } from "@agentclientprotocol/sdk";
 import {CodexEventHandler} from "./CodexEventHandler";
-import {CodexElicitationHandler} from "./CodexElicitationHandler";
 import {CodexAuthMethods, type CodexAuthRequest} from "./CodexAuthMethod";
 import {CodexAcpClient, type SessionMetadata, type SessionMetadataWithThread} from "./CodexAcpClient";
 import type {McpStartupResult} from "./CodexAppServerClient";
@@ -716,9 +715,8 @@ export class CodexAcpServer implements acp.Agent {
         sessionState.lastTokenUsage = null;
 
         try {
-            const eventHandler = new CodexEventHandler(this.connection, sessionState);
-            const elicitationHandler = new CodexElicitationHandler(this.connection, sessionState);
-            this.codexAcpClient.subscribeToSessionEvents(params.sessionId, eventHandler, elicitationHandler);
+            const sessionHandler = new CodexEventHandler(this.connection, sessionState);
+            this.codexAcpClient.subscribeSession(params.sessionId, sessionHandler);
 
             if (await this.availableCommands.tryHandle(params.prompt, sessionState)) {
                 logger.log("Prompt handled by a command");
@@ -767,7 +765,7 @@ export class CodexAcpServer implements acp.Agent {
                 };
             }
 
-            const error = eventHandler.getFailure()
+            const error = sessionHandler.getFailure()
             if (error) {
                 // noinspection ExceptionCaughtLocallyJS
                 throw error;
