+++ src/__tests__/CodexACPAgent/CodexAcpClient.test.ts
@@ -326,17 +326,17 @@ describe('ACP server test', { timeout: 40_000 }, () => {
             { method: "item/agentMessage/delta", params: { threadId: "string", turnId: "string", itemId: "string", delta: "ll", }},
             { method: "item/agentMessage/delta", params: { threadId: "string", turnId: "string", itemId: "string", delta: "o!", }},
         ];
-        function onServerNotification(callback: (event: ServerNotification) => Promise<void>){
+        function subscribeSession(_sessionId: string, handler: { handleNotification(event: ServerNotification): Promise<void> }){
             for (const notification of serverNotifications) {
-                void callback(notification);
+                void handler.handleNotification(notification);
             }
             return { dispose() {} };
         }
-        return onServerNotification;
+        return subscribeSession;
     }
 
     it('should map events from dump', async () => {
-        fixture.getCodexAppServerClient().onServerNotification = loadNotifications();
+        fixture.getCodexAppServerClient().subscribeSession = loadNotifications();
 
         const codexAcpAgent = fixture.getCodexAcpAgent();
